Ex11.py
Write a function that finds the magnitude of a given vector. The magnitude of a vector is the square root of sum of squares of all the components of the vector. Your input for this function will be a (vector with x,y,z components) 1 dimensional list containing 3 integers. For example if the input list is:

vector = [2, 3, -4]
Then you should return the magnitude (as a floating point number) of this vector:
5.385164807134504